% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core.R
\name{scroll}
\alias{scroll}
\title{Scroll}
\usage{
scroll(id, block = c("start", "center", "end", "nearest"))
}
\arguments{
\item{id}{Id of element to scroll to.}

\item{block}{Defines vertical alignment.}
}
\description{
Scroll to a certain element in Shiny.
}
